+++ tests/executors/test_kubernetes_executor.py
@@ -29,13 +29,13 @@ try:
     from kubernetes.client.rest import ApiException
     from airflow import configuration
     from airflow.configuration import conf
-    from airflow.contrib.executors.kubernetes_executor import AirflowKubernetesScheduler
-    from airflow.contrib.executors.kubernetes_executor import KubernetesExecutor
-    from airflow.contrib.executors.kubernetes_executor import KubeConfig
-    from airflow.contrib.executors.kubernetes_executor import KubernetesExecutorConfig
-    from airflow.contrib.kubernetes.worker_configuration import WorkerConfiguration
+    from airflow.executors.kubernetes_executor import AirflowKubernetesScheduler
+    from airflow.executors.kubernetes_executor import KubernetesExecutor
+    from airflow.executors.kubernetes_executor import KubeConfig
+    from airflow.executors.kubernetes_executor import KubernetesExecutorConfig
+    from airflow.kubernetes.worker_configuration import WorkerConfiguration
     from airflow.exceptions import AirflowConfigException
-    from airflow.contrib.kubernetes.secret import Secret
+    from airflow.kubernetes.secret import Secret
 except ImportError:
     AirflowKubernetesScheduler = None  # type: ignore
